% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_bio.R
\docType{data}
\name{ca_bio}
\alias{ca_bio}
\title{Bioclimatic variables of continental Central America}
\format{
A stars object with the bioclimatic variables of continental Central America (EPSG:4326)
}
\source{
The data was obtained using the R package geodata. The data source is WorldClim version 2.1 \url{https://www.worldclim.org/}
}
\usage{
ca_bio
}
\description{
Bioclimatic variables of continental Central America, resolution 1km (0.5 minutes of a degree)
}
\keyword{datasets}
