% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_elevation.R
\docType{data}
\name{ca_elevation}
\alias{ca_elevation}
\title{Elevation data of continental Central America}
\format{
A stars object with elevation data of continental Central America (EPSG:4326)
}
\source{
The data was obtained using the R package geodata. The main data source is Shuttle Radar Topography Mission (SRTM)
}
\usage{
ca_elevation
}
\description{
Elevation data of continental Central America, resolution 1km (30 arc seconds)
}
\keyword{datasets}
