% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_outline.R
\docType{data}
\name{ca_outline}
\alias{ca_outline}
\title{Outline of countries of Central America}
\format{
An sf object with the outline of countries of Central America (EPSG:4326)
}
\source{
\url{https://gadm.org/index.html}
}
\usage{
ca_outline
}
\description{
Outline of countries of Central America
}
\keyword{datasets}
