% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_outline_c.R
\docType{data}
\name{ca_outline_c}
\alias{ca_outline_c}
\title{Outline of countries of Central America (without islands)}
\format{
An sf object with the outline of countries of Central America (EPSG:4326)
}
\source{
\url{https://www.naturalearthdata.com}
}
\usage{
ca_outline_c
}
\description{
Outline of countries of Central America (continent, without islands)
}
\keyword{datasets}
