% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_outline_d.R
\docType{data}
\name{ca_outline_d}
\alias{ca_outline_d}
\title{Outline of Central America (without islands)}
\format{
An sf object with the outline of Central America (EPSG:4326)
}
\source{
\url{https://www.naturalearthdata.com}
}
\usage{
ca_outline_d
}
\description{
Outline of Central America (continent, without islands)
}
\keyword{datasets}
