% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_prec.R
\docType{data}
\name{ca_prec}
\alias{ca_prec}
\title{Total precipitation data of continental Central America}
\format{
A stars object with the total precipitation data of continental Central America (EPSG:4326)
}
\source{
The data was obtained using the R package geodata. The data source is WorldClim version 2.1 \url{https://www.worldclim.org/}
}
\usage{
ca_prec
}
\description{
Total precipitation data of continental Central America (mm), resolution 1km (0.5 minutes of a degree)
}
\keyword{datasets}
