% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_srad.R
\docType{data}
\name{ca_srad}
\alias{ca_srad}
\title{Incident solar radiation data of continental Central America}
\format{
A stars object with the incident solar radiation data of continental Central America (EPSG:4326)
}
\source{
The data was obtained using the R package geodata. The data source is WorldClim version 2.1 \url{https://www.worldclim.org/}
}
\usage{
ca_srad
}
\description{
Incident solar radiation data of continental Central America (kJ m^-2 day^-1), resolution 1km (0.5 minutes of a degree)
}
\keyword{datasets}
