% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_tavg.R
\docType{data}
\name{ca_tavg}
\alias{ca_tavg}
\title{Average temperature data of continental Central America}
\format{
A stars object with the average temperature data of continental central America (EPSG:4326)
}
\source{
The data was obtained using the R package geodata. The data source is WorldClim version 2.1 \url{https://www.worldclim.org/}
}
\usage{
ca_tavg
}
\description{
Average temperature data of continental Central America (°C), resolution 1km (0.5 minutes of a degree)
}
\keyword{datasets}
