% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_tmax.R
\docType{data}
\name{ca_tmax}
\alias{ca_tmax}
\title{Maximum temperature data of continental Central America}
\format{
A stars object with the maximum temperature data of continental Central America (EPSG:4326)
}
\source{
The data was obtained using the R package geodata. The data source is WorldClim version 2.1 \url{https://www.worldclim.org/}
}
\usage{
ca_tmax
}
\description{
Maximum temperature data of continental Central America (°C), resolution 1km (0.5 minutes of a degree)
}
\keyword{datasets}
