% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_wind.R
\docType{data}
\name{ca_wind}
\alias{ca_wind}
\title{Wind speed data of continental Central America}
\format{
A stars object with wind speed data of continental Central America (EPSG:4326)
}
\source{
The data was obtained using the R package geodata. The data source is WorldClim version 2.1 \url{https://www.worldclim.org/}
}
\usage{
ca_wind
}
\description{
Wind speed (2m above the ground) data of continental Central America (m s^-1), resolution 1km (0.5 minutes of a degree)
}
\keyword{datasets}
