\name{gpuReset}
\alias{gpuReset}
\title{GPU Device Reset}

\description{
Resets the GPU device and releases allocated device memory 
}

\usage{
	gpuReset() 
}


\value{No value is returned}

\author{Marius Appel \email{marius.appel@uni-muenster.de}}

\examples{
gpuReset()
}

\keyword{cuda}
\keyword{gpu}
