% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cageo-package.R
\docType{package}
\name{cageo-package}
\alias{cageo}
\alias{cageo-package}
\title{cageo: cageo: Geospatial data of Central America}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Contains geospatial data of Central America
}
\seealso{
Useful links:
\itemize{
  \item \url{https://manuelspinola.github.io/cageo/}
}

}
\author{
\strong{Maintainer}: Manuel Spínola \email{mspinola10@gmail.com} (\href{https://orcid.org/-9749}{ORCID})

}
\keyword{internal}
